feat(layout): make footer
