+++ src/app/page.tsx
@@ -1,23 +1,7 @@
-"use client"
-
-import { AnimatePresence } from "framer-motion"
-import { useEffect, useState } from "react"
-
-import LoadingScreen from "@/components/loading-screen"
-
 export default function Home() {
-  const [loading, setLoading] = useState(true)
-
-  useEffect(() => {
-    const timer = setTimeout(() => setLoading(false), 2000)
-    return () => clearTimeout(timer)
-  }, [])
-
   return (
     <>
-      <AnimatePresence>{loading && <LoadingScreen />}</AnimatePresence>
-
-      {!loading && <div className="">Home</div>}
+      <div className="">Home</div>
     </>
   )
 }
