+++ src/components/layout/footer/index.tsx
@@ -1,0 +1,31 @@
+import { siteConfig } from "@/config/site"
+
+export default function Footer() {
+  return (
+    <div className="py-6 md:px-8 md:py-0">
+      <div className="container flex flex-col items-center justify-between gap-4 md:h-24 md:flex-row">
+        <p className="text-bone text-balance text-center text-sm leading-loose md:text-left">
+          Built by{" "}
+          <a
+            href={siteConfig.links.github}
+            target="_blank"
+            rel="noreferrer"
+            className="font-medium underline underline-offset-4"
+          >
+            Nopnapatn
+          </a>
+          . This exquisite work of code is readily accessible for exploration on{" "}
+          <a
+            href={siteConfig.links.github}
+            target="_blank"
+            rel="noreferrer"
+            className="font-medium underline underline-offset-4"
+          >
+            GitHub
+          </a>
+          .
+        </p>
+      </div>
+    </div>
+  )
+}
